feat(network): refactored directories
