+++ src/app/networkDiscovery/components/DiscoveriesTable/DiscoveriesFilterAccordion/DiscoveriesFilterAccordion.test.tsx
