+++ src/app/networkDiscovery/components/DiscoveriesTable/DiscoveriesFilterAccordion/DiscoveriesFilterAccordion.tsx
