+++ src/app/networkDiscovery/components/DiscoveryAddForm/DiscoveryAddFormFields/DiscoveryAddFormFields.test.tsx
