+++ src/app/networkDiscovery/components/DiscoveryAddForm/DiscoveryAddFormFields/DiscoveryAddFormFields.tsx
