+++ src/app/networkDiscovery/components/DiscoveryAddForm/DiscoveryAddFormFields/index.ts
