+++ src/app/networkDiscovery/components/DiscoveryAddForm/types.ts
