+++ src/app/networkDiscovery/components/DiscoveryDeleteForm/DiscoveryDeleteForm.test.tsx
