+++ src/app/networkDiscovery/components/DiscoveryDeleteForm/DiscoveryDeleteForm.tsx
