+++ src/app/networkDiscovery/components/DiscoveryDeleteForm/index.ts
