+++ src/app/networkDiscovery/components/NetworkDiscoveryHeader/ClearAllForm/ClearAllForm.test.tsx
