+++ src/app/networkDiscovery/components/NetworkDiscoveryHeader/ClearAllForm/ClearAllForm.tsx
