+++ src/app/networkDiscovery/components/NetworkDiscoveryHeader/ClearAllForm/index.ts
