+++ src/app/networkDiscovery/components/NetworkDiscoveryHeader/NetworkDiscoveryHeader.test.tsx
