+++ src/app/networkDiscovery/components/NetworkDiscoveryHeader/NetworkDiscoveryHeader.tsx
@@ -2,7 +2,7 @@ import { Button } from "@canonical/react-components";
 import pluralize from "pluralize";
 import { useLocation, Link } from "react-router";
 
-import { NetworkDiscoverySidePanelViews } from "../constants";
+import { NetworkDiscoverySidePanelViews } from "../../views/constants";
 
 import { useNetworkDiscoveries } from "@/app/api/query/networkDiscovery";
 import SectionHeader from "@/app/base/components/SectionHeader";
