--- src/app/networkDiscovery/components/NetworkForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./NetworkForm";
